Move foreign func name resolution
Move foreign function name resolution into its own class. In the future this class will need to use the classfile API to resolve function names instead of using reflection, this should make that easier.

+++ src/main/java/com/pentlander/sasquach/ast/ModuleScopedId.java
@@ -11,6 +11,6 @@ public record ModuleScopedId(Id moduleId, Id id) implements Identifier {
 
   @Override
   public Single range() {
-    return (Single) moduleId.range().join(id.range());
+    return  moduleId.range().join(id.range());
   }
 }
